add validation

+++ src/archive/archive.controller.ts
@@ -1,6 +1,7 @@
-import { Controller, Delete, Get, Param, Post } from '@nestjs/common';
+import { Controller, Delete, Get, HttpException, Param, Post } from '@nestjs/common';
 import { Note } from 'src/helpers/types';
 import { NotesService } from 'src/notes/notes.service';
+import { checkId } from 'src/notes/validations';
 import { ArchiveService } from './archive.service';
 
 @Controller('archive')
@@ -12,6 +13,7 @@ export class ArchiveController {
 
   @Post(':id')
   toArchive(@Param('id') id: string) : Note[] {
+    if(!checkId(+id)) throw new HttpException("Not founded", 404);
     ArchiveService.add(NotesService.getOne(+id));
     NotesService.delete(+id);
     return ArchiveService.getAll();
@@ -19,6 +21,7 @@ export class ArchiveController {
 
   @Delete(':id')
   fromArchive(@Param('id') id: string) : Note[]{
+    if(!checkId(+id)) throw new HttpException("Not founded", 404);
     NotesService.addFromArchive(ArchiveService.remove(+id));
     return ArchiveService.getAll();
   }
